Move is_sol_keyword to stylus-core
