+++ stylus-sdk/Cargo.toml
@@ -18,10 +18,6 @@ clap.workspace = true
 derivative.workspace = true
 hex = { workspace = true, default-features = false, features = ["alloc"] }
 keccak-const.workspace = true
-lazy_static.workspace = true
-
-# export-abi
-regex = { workspace = true, optional = true }
 
 # local deps
 mini-alloc = { workspace = true, optional = true }
@@ -44,7 +40,7 @@ features = ["default", "docs", "debug", "export-abi", "stylus-test"]
 
 [features]
 default = ["mini-alloc"]
-export-abi = ["debug", "regex", "stylus-proc/export-abi", "alloy-primitives/tiny-keccak"]
+export-abi = ["debug", "stylus-proc/export-abi", "alloy-primitives/tiny-keccak"]
 debug = []
 docs = []
 hostio = []
